Don't crash on missing entities data
I'm not sure if the response this code is meant to handle can legitimately
either include or not include the data it is trying to parse, or whether the
format has changed (and the data is never included, or never in that format)
and this tool simply hasn't kept up. Either way, not crashing seems better than
crashing.